requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=2.0
django-cors-headers>=3.15
python-telegram-bot>=20.5
requests>=2.31
python-dotenv>=1.0.0
